Splitted the callback and processor of Calibre
